requirements.txt
numpy==1.22.2
pandas==1.4.0
Pillow==9.2.0
praw==7.6.0
requests==2.28.1
scipy==1.8.1
